vc4: move all vc4_bo_create() to kernel

+++ ucore/src/kern-ucore/arch/arm/mach-raspberrypi/vc4/vc4_bo.c
@@ -106,7 +106,7 @@ int vc4_mmap_bo_ioctl(struct device *dev, void *data)
 		return -E_INVAL;
 	}
 
-	args->offset = (uint32_t)bo->vaddr;
+	args->offset = (uintptr_t)bo->vaddr;
 
 	return 0;
 }
